# crates/service/openapi/fluid-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crystalline Fluid API",
    "description": "The JSON API the Crystalline daemon mounts at `/api/v1` on its HTTP endpoint (127.0.0.1:7411 by default), the surface the Fluid UI talks to. Reading is open to any signed-in viewer; writing content needs an editor account and the `If-Match` token of the version being replaced, and account management needs an admin.\n\nEvery path but `/auth/login`, `/auth/logout`, `/auth/me` and `/auth/setup` is closed by default: a request that carries no identity is answered 401 ahead of routing, so an unauthenticated caller never learns which paths exist. Every failure is an RFC 9457 problem detail sent as `application/problem+json`.\n\nThe payloads marked as generic objects are the engine's own JSON, passed through unchanged so this API and the MCP tools stay one source of truth; each carries an example of the shape it answers with.",
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/settings/github": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "The GitHub connection, and the device flow's poll.",
        "description": "Admin only. A pure read, served even on a read-only instance. Also the device flow's poll: while one runs, `pending` carries the short code and its URL; once it ends, `pending` goes null and a failed flow's reason is reported in `error` on exactly one read, then cleared.",
        "operationId": "get_github_settings",
        "responses": {
          "200": {
            "description": "The connection. No token material.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "No identity.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not an admin, or the trusted-header identity names a disabled account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Forget the stored credential.",
        "description": "Admin only, and refused on a read-only instance. Idempotent: disconnecting an instance that holds no credential succeeds and answers `connected: false` rather than 404. `github.enabled` is left alone - turning the feature off is a separate intent.",
        "operationId": "disconnect_github",
        "responses": {
          "200": {
            "description": "The connection as it now stands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "No identity.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not an admin, a cookie session did not echo its CSRF token, this instance is read-only, or the trusted-header identity names a disabled account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "The credential comes from `CRYSTALLINE_GITHUB_TOKEN`: only the environment that set it can retire it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/github/connect": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Start a device-code sign-in.",
        "description": "Admin only, and refused on a read-only instance. Answers 202 with the short code to confirm in a browser; the flow itself runs in the background and its outcome is read from `GET /settings/github`. A second call while one flow runs reports that same flow rather than starting another. Connecting is the intent to use the feature, so this turns `github.enabled` on.",
        "operationId": "connect_github_device",
        "responses": {
          "202": {
            "description": "The flow started (or one was already running): `pending` carries the code to confirm. Poll `GET /settings/github` for the outcome.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "No identity.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not an admin, a cookie session did not echo its CSRF token, this instance is read-only, or the trusted-header identity names a disabled account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "This machine's identity is fixed by `CRYSTALLINE_GITHUB_TOKEN`, so no sign-in can be started here, or GitHub refused to start the flow.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/github/token": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Connect with a personal access token.",
        "description": "Admin only, and refused on a read-only instance. The token is validated against GitHub before it is stored, and is never echoed: the answer is the same token-material-free status the GET returns. Connecting is the intent to use the feature, so this turns `github.enabled` on.",
        "operationId": "connect_github_token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The connection as it now stands.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GithubStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "The body is not JSON.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "No identity.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "The caller is not an admin, a cookie session did not echo its CSRF token, this instance is read-only, or the trusted-header identity names a disabled account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "415": {
            "description": "The body is not `application/json`.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "The body is JSON but not a token, the token is empty, GitHub refused it, or this machine's identity is fixed by `CRYSTALLINE_GITHUB_TOKEN` and no token may be stored here.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GithubPendingView": {
        "type": "object",
        "description": "The half of a running device flow a browser has to show: the short code the user types in, where they type it, and how long the code stays valid.",
        "required": [
          "user_code",
          "verification_url",
          "expires_in_secs"
        ],
        "properties": {
          "expires_in_secs": {
            "type": "integer",
            "format": "int64",
            "description": "How many seconds from the flow's start the code stays valid.",
            "example": 900,
            "minimum": 0
          },
          "user_code": {
            "type": "string",
            "description": "The short code the user confirms at `verification_url`.",
            "example": "ABCD-1234"
          },
          "verification_url": {
            "type": "string",
            "description": "Where the user confirms the code.",
            "example": "https://github.com/login/device"
          }
        }
      },
      "GithubStatusResponse": {
        "type": "object",
        "description": "The GitHub connection as the settings screen renders and polls it. No token material, ever: only whether the feature is on, whether a credential is on file, whose it is and where it lives.",
        "required": [
          "enabled",
          "connected"
        ],
        "properties": {
          "connected": {
            "type": "boolean",
            "description": "Whether a credential is on file for this instance.",
            "example": true
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether `github.enabled` is on: team tools and origin polling.",
            "example": true
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "A device flow's failure (expired, denied), reported on exactly one\nstatus read after the flow ends, then cleared."
          },
          "pending": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GithubPendingView",
                "description": "A device flow waiting for the browser side. Poll this route until it\ngoes null; the flow either connected or set `error`."
              }
            ]
          },
          "token_store": {
            "type": [
              "string",
              "null"
            ],
            "description": "keyring | file | environment; null when disconnected.",
            "example": "keyring"
          },
          "user": {
            "type": [
              "string",
              "null"
            ],
            "description": "The account login, when connected.",
            "example": "octo"
          }
        }
      },
...
      "ProblemDetail": {
        "type": "object",
        "description": "An RFC 9457 problem detail, sent as `application/problem+json`. Every failure on this surface has this shape, so a client can branch on `status` alone.",
        "required": [
          "type",
          "status",
          "title",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "The specific occurrence, safe to show to the caller.",
            "example": "no engram 'ghost' in domain 'eng'"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status, mirrored into the body so a client that only has the\nparsed payload can still branch on it.",
            "example": 404,
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "A short, stable, human-readable summary of the problem type.",
            "example": "not found"
          },
          "token_required": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "An RFC 9457 extension member, present only on the `403` of\n`POST /auth/setup` and only when this instance holds a setup token: the\nfirst-run wizard renders its token field on this member rather than on\nthe detail prose, so an instance that has no token to enter (the\nloopback bind generates none) never causes a dead-end input to be\ndrawn. Absent from every other problem document on this surface.",
            "example": true
          },
          "type": {
            "type": "string",
            "description": "The problem type URI. Always `about:blank`: `status` and `title` carry\nthe classification, and this surface publishes no per-problem pages to\npoint at.",
            "example": "about:blank"
          }
        }
      },
...
      "TokenBody": {
        "type": "object",
        "description": "A GitHub personal access token to connect with. Write-only: no response on this surface ever echoes it, and the status shape carries only where the credential lives and whose it is.",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "A GitHub personal access token. Write-only: no response ever echoes it.",
            "example": "ghp_xxxxxxxxxxxxxxxxxxxx",
            "writeOnly": true
          }
        }
      },
...
    }
  },
...
}